+++ packages/react/src/components/Avatar/AvatarContent.tsx
@@ -3,6 +3,8 @@ import { useState } from 'react';
 import { AvatarProps } from './Avatar';
 import * as Styled from './Avatar.style';
 import { nameToAbbreviation } from './utils';
+import { getPxFromRem } from '../../utils/getPxFromRem';
+import useTheme from '../../hooks/useTheme';
 
 const baseTestId = 'avatar';
 
@@ -11,11 +13,21 @@ const AvatarContent = ({
   source,
   icon,
   inactive,
-  size,
+  size = 32,
   testID = '',
 }: AvatarProps) => {
   const abbreviation = name && nameToAbbreviation(name);
   const [imageError, setImageError] = useState(false);
+  const theme = useTheme();
+
+  const getTypographyVariant = (size: number) => {
+    const mediumBreakpoint = getPxFromRem(theme.SpacingBase48);
+    const largeBreakPoint = getPxFromRem(theme.SpacingBase80);
+
+    if (size <= mediumBreakpoint) return 'heading6';
+    if (size <= largeBreakPoint) return 'heading5';
+    return 'heading1';
+  };
 
   if (source && !imageError) {
     return (
@@ -31,7 +43,7 @@ const AvatarContent = ({
 
   if (abbreviation) {
     return (
-      <Styled.Text inactive={inactive} variant={'heading6'}>
+      <Styled.Text inactive={inactive} variant={getTypographyVariant(size)}>
         {abbreviation}
       </Styled.Text>
     );