+++ packages/react/src/components/Link/Link.style.ts
@@ -1,7 +1,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import styled, { FlattenSimpleInterpolation, css } from 'styled-components';
 import { ExtendedTheme, Theme } from '../../types';
-import { TypographyVariant } from '../Typography';
+import { TypographyVariant, getColorStyles } from '../Typography';
 import { ThemeDesignTokens } from '../../theme/ThemeProvider';
 import { LinkProps } from './Link';
 
@@ -56,9 +56,13 @@ export const getIconSize = (theme: ThemeDesignTokens): Record<TypographyVariant,
   `,
 });
 
-export const Link = styled.a`
+type LinkType = Omit<LinkProps, 'variant'>;
+
+export const Link = styled.a<LinkType>`
   text-decoration: underline;
   cursor: pointer;
+  text-decoration: underline
+    ${({ theme, mode = 'light', color = 'primary' }: ExtendedTheme<LinkType>) => getColorStyles(theme)[mode][color]};
 
   ${({ theme }: Theme) => css`
     color: ${theme.LinkColor};