+++ packages/react/src/components/Typography/index.ts
@@ -1,2 +1,3 @@
 export { default, TypographyProps } from './Typography';
 export { TypographyVariant } from './Typography.types';
+export { getColorStyles } from './Typography.style';